backend/requirements.txt
anyio==3.6.2
asttokens==2.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.2.0
category-encoders==2.5.1.post0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.5
cssselect==1.1.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.8
distro==1.9.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fastapi==0.94.1
filelock==3.13.1
fonttools==4.37.4
google-api-core==2.10.2
google-api-python-client==2.64.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
h11==0.14.0
httpcore==1.0.4
httplib2==0.20.4
httptools==0.5.0
httpx==0.27.0
idna==3.4
imbalanced-learn==0.10.0
imblearn==0.0
importlib-metadata==5.0.0
install==1.3.5
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
joblib==1.2.0
jupyter-client==7.4.5
jupyter-core==5.0.0
kiwisolver==1.4.4
lxml==4.6.3
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mlxtend==0.21.0
mysql-connector-python==8.0.28
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.23.3
oauthlib==3.2.1
openai==1.12.0
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==4.1.0
prettytable==3.5.0
prompt-toolkit==3.0.32
protobuf==4.21.7
psutil==5.9.4
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.6
Pygments==2.13.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyquery==1.4.3
pyspark==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.2.1
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
regex==2023.8.8
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
selenium==3.141.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.1
starlette==0.26.1
statsmodels==0.13.5
threadpoolctl==3.1.0
torch==1.13.0
tornado==6.2
tqdm==4.66.1
traitlets==5.5.0
typing-extensions==4.10.0
ujson==5.7.0
uritemplate==4.1.1
urllib3==1.26.5
uvicorn==0.21.0
virtualenv==20.25.0
watchfiles==0.18.1
wcwidth==0.2.5
websockets==10.4
xgboost==1.7.2
zipp==3.10.0
